+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1 (unreleased)
 
+- Added support for international phone numbers
 - Added support for MAC addresses
 
 ## 0.1.0 (2022-08-21)